Isolate retrieval module
